convert matrix to float32 in np svd

+++ llama/flex_model.py
@@ -153,7 +153,7 @@ def cupy_decompose_matrix(matrix, target_dim):
 
 def numpy_decompose_matrix(matrix, target_dim):
     # Assuming matrix is a PyTorch tensor on a GPU, move it to CPU and convert to a NumPy array.
-    matrix_np = matrix.cpu().numpy()
+    matrix_np = matrix.float().cpu().numpy()
 
     # Perform SVD using NumPy.
     U_np, S_np, V_np = np.linalg.svd(matrix_np, full_matrices=True)
